Refactor for patchwork file, split into two
