+++ src/game_state.rs
@@ -2,7 +2,7 @@ pub mod block;
 pub mod patchwork;
 pub mod player;
 
+use super::map;
 use super::messenger;
 use super::packet;
 use super::packet_processor;
-use super::server;
